Move domain decomposition functionality to a separate class in a separate file
